Refactor BaseInvoiceForm component and update tax rates
