+++ src/lib/constants/settings.ts
@@ -1,1 +1,5 @@
+// todo: store these values on the server
 export const IVA_RATE = 12;
+export const ISD_TAX_RATE = 3.5;
+export const FODINFA_TAX_RATE = 0.5;
+export const INSURANCE_RATE = 1;
